refactor(backend): isolate skill discovery domain
