+++ scripts/tasks/supported-platform-structure-assets.json
@@ -195,7 +195,7 @@
     ],
     [
       "src-tauri/src/services/skill.rs",
-      "b85c42c8ff3ed851685cdcef1af68e4f519cd17e54fdda3b2299ae784ac4ee13"
+      "38d22a4ed94483302e11947cdc0b542d5ad1e31678bbf7280734045128b73825"
     ],
     [
       "src-tauri/src/services/subscription.rs",
